Converted all Activities to Fragment to fix compatilibility problem + added custom toolbar

+++ app/src/main/java/com/lubenard/ft_hangouts/DbManager.java
@@ -97,7 +97,8 @@ public class DbManager extends SQLiteOpenHelper {
      */
     public void deleteContact(int contactId)
     {
-        writableDB.delete(contactsTable,contactsTableId + "=?", new String[]{String.valueOf(contactId)});
+        if (contactId > 0)
+            writableDB.delete(contactsTable,contactsTableId + "=?", new String[]{String.valueOf(contactId)});
     }
 
     /**
